trace: the wire could not carry a u128, and every unit test missed it
Caught by the first live run, not by the suite. `ForeignTraceRecord.ts_ms` was
`u128`; the record crosses the webview bridge as a `serde_json::Value`, whose
number type is u64/i64/f64. Deserialization failed with "u128 is not supported"
and the bridge discarded the WHOLE batch — which reached the trace as
`js_event_ignored` and looked exactly like a foreign layer with nothing to say.
That is the failure mode this contract spends most of its rules preventing, and
it shipped inside the mechanism meant to prevent it.

Every existing contract test built the struct in Rust and therefore skipped the
wire, so a test that constructs its own input was testing the constructor. The
new one decodes from a Value, the way the bridge does, and also decodes the
minimum an emitter may legally send.

u64 millis run to the year 584 million; the record widens to the plane's u128 at
the boundary, where it is no longer JSON.

Also guarded: the replay script's reseed wiring. A retained replay needs a
re-attach with daemon-held history, which a fresh sandbox never produces, so
that arm has no live proof yet — the wipe/refill marks and the reseed-tagged
capture are asserted against the generated script instead of assumed.

+++ crates/yggterm-shell/src/shell/viewport.rs
@@ -9683,8 +9683,8 @@ fn TerminalCanvas(
                                         .map(|record| record.ts_ms)
                                         .max()
                                         .unwrap_or_default();
-                                    let flush_lag_ms = u128::from(current_millis())
-                                        .saturating_sub(newest_ts_ms);
+                                    let flush_lag_ms =
+                                        current_millis().saturating_sub(newest_ts_ms);
                                     let submitted = records.len();
                                     let outcome =
                                         append_foreign_trace_batch(&trace_home, records);
